Benchmarking by means of the Cholette method

Cholette method is based on a benchmarking methodology developed at Statistics Canada. It is a generalized method relying on the principle of movement preservation that encompasses other benchmarking methods. The Denton method (both the AFD and PFD variants), as well as the naive pro-rating method, emerge as particular cases of the Cholette method. This method has been widely used for the purpose of benchmarking seasonally adjusted series among others.

Usage

cholette(
  s,
  t,
  rho = 1,
  lambda = 1,
  bias = c("None", "Additive", "Multiplicative"),
  conversion = c("Sum", "Average", "Last", "First", "UserDefined"),
  obsposition = 1L
)

Arguments

s

Preliminary series. Mandatory. It must be the same class as t.

t

Aggregation constraint. Mandatory. It must be either an object of class ts or a numeric vector.

rho

Numeric. Smoothing parameter whose value should be between 0 and 1. See vignette for more information on the choice of the rho parameter.

lambda

Numeric. Adjustment model parameter. Typically, lambda = 1 for proportional benchmarking; lambda = 0 for additive benchmarking; and lambda = 0.5 with rho = 0 for the naive pro-rating method. See vignette for more information on the choice of the lambda parameter.

bias

Character. Bias correction factor. No systematic bias is considered by default. See vignette for more details.

conversion

Conversion rule. Usually "Sum" or "Average". Sum by default.

obsposition

Position of the observation in the aggregated period (only used with "UserDefined" conversion).

Value

The benchmarked series is returned

References

Quenneville, B., Fortier S., Chen Z.-G., Latendresse E. (2006). Recent Developments in Benchmarking to Annual Totals in X12-ARIMA and at Statistics Canada. Statistics Canada, Working paper of the Time Series Research and Analysis Centre.

See also

Examples

ym_true <- rjd3toolkit::Retail$RetailSalesTotal
yq_true <- rjd3toolkit::aggregate(ym_true, 4)
Y_full <- rjd3toolkit::aggregate(ym_true, 1)

Y <- window(Y_full, end = c(2009,1)) # say no benchmark yet for the year 2010
xm <- ym_true + rnorm(n = length(ym_true), mean = -5000, sd = 10000)
xq <- rjd3toolkit::aggregate(xm, 4)

# Proportional benchmarking with a bias and some recommended value of rho for
# monthly and quarterly series respectively (see vignette)
cholette(s = xm, t = Y, rho = 0.9, lambda = 1, bias = "Multiplicative")
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 129476.7 136648.0 153798.2 128168.5 149945.8 149472.7 149966.0 147154.3
#> 1993 118020.0 129407.5 153633.1 165769.0 155842.0 169681.3 170375.9 152449.4
#> 1994 146559.3 139242.4 160567.4 183125.2 180109.3 183750.0 192356.8 185963.3
#> 1995 168026.2 154400.0 174018.3 174208.7 185658.3 193739.9 208033.8 191101.5
#> 1996 162077.2 169321.0 189949.7 190290.1 209911.8 172315.0 223221.7 199740.3
#> 1997 178407.9 166368.5 194661.6 211168.9 214353.3 211887.5 192828.3 198328.2
#> 1998 192373.0 188988.4 205934.8 212688.5 218588.5 238188.1 228130.9 220973.0
#> 1999 189464.9 197564.3 219738.3 224492.1 241137.5 240667.3 230517.9 239912.6
#> 2000 211721.8 226155.6 251018.2 233883.7 256733.7 244434.3 244521.9 259355.6
#> 2001 212618.4 225048.9 246313.9 250789.0 285228.4 259372.2 229365.1 263538.2
#> 2002 233256.4 228672.4 258283.0 265195.1 271823.2 258052.0 272455.7 270059.7
#> 2003 247050.8 223976.0 260606.8 273415.2 280184.6 267081.4 268199.8 301069.9
#> 2004 255135.4 235891.2 301436.7 286960.0 296236.9 277132.2 288578.9 304494.1
#> 2005 272752.9 247398.5 318870.3 301975.6 317625.4 303019.7 312937.7 314184.6
#> 2006 279289.3 298280.5 339735.9 327479.6 351140.4 328406.1 314512.0 337114.9
#> 2007 305920.3 290012.6 348811.2 312898.7 370067.4 340137.8 336472.8 351583.8
#> 2008 304973.1 309272.6 344529.3 343162.1 364451.8 336798.2 360204.1 344387.4
#> 2009 270136.1 245900.9 320382.2 283379.8 307724.1 318024.6 299791.2 324293.8
#> 2010 293445.2 276841.6 326776.5 324356.5 333414.3 316459.3 341257.4 333610.3
#>           Sep      Oct      Nov      Dec
#> 1992 155051.8 178175.0 138725.9 199133.0
#> 1993 168943.6 185889.4 171506.9 200729.8
#> 1994 175437.6 160318.1 182532.3 220059.2
#> 1995 185453.5 182533.0 171299.7 234031.1
#> 1996 190654.3 212495.5 196704.3 249984.2
#> 1997 205429.8 225088.4 210845.4 264635.3
#> 1998 210789.3 210844.7 225354.1 234251.7
#> 1999 242913.1 226281.0 248558.1 307308.8
#> 2000 244470.9 253159.1 266928.1 296373.1
#> 2001 241756.2 280943.9 262574.8 310175.9
#> 2002 258019.5 263857.4 252735.7 301911.9
#> 2003 269775.2 280398.1 282511.8 313884.6
#> 2004 289701.7 288167.2 291494.5 365201.3
#> 2005 308033.9 303191.3 321199.4 375501.9
#> 2006 332923.8 301267.4 309445.3 360541.0
#> 2007 305348.7 315480.5 329485.7 399578.4
#> 2008 313089.9 289414.9 302889.3 339760.2
#> 2009 292279.6 310125.5 312314.6 354118.6
#> 2010 309711.9 329990.6 320849.7 399381.2
cholette(s = xq, t = Y, rho = 0.729, lambda = 1, bias = "Multiplicative")
#>           Qtr1      Qtr2      Qtr3      Qtr4
#> 1992  419922.5  427571.6  452171.5  516050.4
#> 1993  401071.9  491346.5  491779.3  558050.4
#> 1994  446249.3  546841.9  553681.0  563248.7
#> 1995  496256.2  553656.2  584677.7  587914.0
#> 1996  521541.3  572617.0  613590.3  658916.4
#> 1997  539447.6  637402.9  596531.7  700620.8
#> 1998  586936.4  669247.4  659893.4  671027.8
#> 1999  607169.0  706856.4  713573.3  780957.3
#> 2000  689086.9  734592.6  747887.5  817189.0
#> 2001  683526.2  795636.6  735198.0  853364.2
#> 2002  720763.4  794924.3  800115.2  818519.1
#> 2003  731115.4  820654.4  839337.8  877046.4
#> 2004  792786.5  860452.3  882681.4  944509.7
#> 2005  839099.7  922693.7  935140.2  999757.4
#> 2006  916979.4 1006663.3  984530.6  971962.7
#> 2007  944660.1 1023404.0  993693.6 1044040.4
#> 2008  959253.8 1044433.7 1017472.2  931773.2
#> 2009  836211.4  909092.3  916442.5  976724.8
#> 2010  897068.5  974223.5  984570.2 1050228.0

# Proportional benchmarking with no bias
xm_no_bias <- ym_true + rnorm(n = length(ym_true), mean = 0, sd = 10000)
cholette(s = xm_no_bias, t = Y, rho = 0.9, lambda = 1)
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 119121.9 140636.2 128740.8 150099.9 148247.9 145534.7 161629.8 166241.0
#> 1993 149904.9 120705.2 167194.6 158848.1 162866.4 167408.6 165965.9 155897.2
#> 1994 138568.1 151359.4 171559.4 167696.7 193743.5 183507.3 161958.2 185693.5
#> 1995 148940.5 154180.2 183716.4 176169.5 167115.2 188427.0 181804.2 190899.1
#> 1996 162317.3 182060.7 180887.0 195746.7 207305.4 199006.0 206977.3 193632.5
#> 1997 178028.1 167633.3 207674.8 214410.3 195695.6 190759.6 217009.5 217664.0
#> 1998 199148.9 185647.4 203818.0 218169.2 240506.6 239223.1 201336.9 210538.1
#> 1999 195572.2 209239.9 225295.8 214145.8 236625.4 240381.7 237454.4 252075.6
#> 2000 225549.7 228830.7 258370.9 247954.3 236478.1 265835.3 241754.4 259520.4
#> 2001 235229.8 221680.2 274386.2 247973.5 254866.8 248999.9 268458.5 265290.6
#> 2002 225729.1 229206.8 268687.7 265534.9 280053.9 250566.3 278611.8 266473.5
#> 2003 258892.2 232202.6 275076.8 250314.2 275423.6 285051.9 280666.3 290311.0
#> 2004 265567.4 258313.8 283259.6 282465.4 307370.2 291475.0 289492.2 290878.8
#> 2005 269678.9 255526.4 306153.7 294504.0 317050.5 332214.7 306830.4 318737.2
#> 2006 282723.0 300777.3 319382.4 308967.9 350245.9 318557.2 316775.1 348378.9
#> 2007 289701.7 283956.4 331158.9 334769.1 358831.8 342583.4 315591.3 357262.1
#> 2008 299270.6 281070.9 344901.8 327810.1 342959.2 345583.0 358716.5 353129.8
#> 2009 275977.1 259670.3 293091.0 301154.1 298003.0 300484.1 324880.5 310834.5
#> 2010 276945.6 298910.6 333661.7 344259.9 338012.8 319123.4 314217.2 330575.8
#>           Sep      Oct      Nov      Dec
#> 1992 133039.6 148719.1 169639.3 204065.8
#> 1993 154673.5 152685.7 163459.1 222638.7
#> 1994 175934.0 172131.1 165290.4 242579.4
#> 1995 198515.0 202920.9 197686.2 232129.9
#> 1996 190460.1 214878.6 198757.4 234635.8
#> 1997 222067.7 206303.2 196732.4 260024.6
#> 1998 196455.1 224981.8 199614.5 267665.2
#> 1999 228596.2 251443.7 223647.5 294077.7
#> 2000 255553.9 233331.7 240487.7 295088.8
#> 2001 231111.2 262632.1 267591.8 289504.6
#> 2002 233135.7 252239.7 270430.5 313652.0
#> 2003 261313.5 267023.4 265122.8 326755.7
#> 2004 280833.4 288518.4 291437.6 350818.3
#> 2005 318435.3 287225.1 322116.1 368218.8
#> 2006 334432.4 309084.6 308714.8 382096.4
#> 2007 318913.1 336661.4 340368.1 396000.6
#> 2008 310801.2 323780.2 317770.7 347138.9
#> 2009 292401.1 313564.0 308950.1 359461.2
#> 2010 294817.5 323822.7 333537.6 408230.0

# Additive benchmarking
cholette(s = xm, t = Y, rho = 0.9, lambda = 0, bias = "Additive")
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 140343.1 144085.5 157969.5 130765.6 149960.3 147963.6 147277.1 143755.5
#> 1993 116944.4 128760.1 152796.1 165069.9 156028.9 169737.2 170727.0 153765.2
#> 1994 148230.6 141099.9 161398.4 182890.0 179900.6 183312.0 191522.6 185346.7
#> 1995 167978.1 154665.4 173936.6 174185.9 185545.4 193610.5 207831.1 191128.7
#> 1996 162516.4 169696.0 190077.6 190417.2 209752.7 172750.4 222822.0 199753.2
#> 1997 179079.3 167408.9 194963.3 211006.9 214095.6 211702.9 193200.3 198539.9
#> 1998 192729.1 189413.6 206054.3 212714.9 218548.7 237942.9 228014.5 220911.3
#> 1999 188996.1 197086.1 219481.7 224352.7 241251.4 240844.5 230681.5 240198.7
#> 2000 213115.4 227297.7 251283.5 234824.2 256613.2 244803.3 244790.1 258808.4
#> 2001 212753.0 224800.8 245857.4 250445.4 285047.8 259319.0 229298.0 263727.1
#> 2002 233929.5 229507.5 258551.8 265244.0 271632.1 258203.3 272135.7 269772.8
#> 2003 247145.1 224420.1 260476.7 273185.9 279966.4 267038.5 268204.7 300954.8
#> 2004 255402.4 236323.1 301362.0 287011.5 296199.5 277319.3 288652.2 304372.4
#> 2005 273395.5 248675.2 318545.8 302070.8 317345.3 303141.3 312826.8 314074.2
#> 2006 280391.8 298932.5 339387.7 327452.4 350602.4 328360.0 314698.7 336876.1
#> 2007 305439.8 289343.4 348574.1 312376.7 370336.3 340115.7 336487.0 351889.8
#> 2008 305844.6 310336.7 345441.6 344004.4 364874.4 337356.5 360082.5 344134.8
#> 2009 267910.3 243600.9 316192.5 280059.0 304260.5 315063.9 298309.4 323710.8
#> 2010 306800.3 293733.9 345551.9 345782.4 357007.4 342486.3 368711.7 362918.4
#>           Sep      Oct      Nov      Dec
#> 1992 150937.7 173158.8 135324.6 194174.8
#> 1993 169756.5 186091.3 172314.4 200257.0
#> 1994 175185.4 160552.2 182032.5 218550.0
#> 1995 185572.1 182706.9 171601.2 233742.1
#> 1996 190858.7 212278.0 196842.7 248900.1
#> 1997 205438.6 224571.9 210726.6 263268.9
#> 1998 210752.7 210751.3 225196.0 234075.8
#> 1999 243255.6 226751.8 248953.2 306702.7
#> 2000 244474.7 252641.1 265776.9 294327.6
#> 2001 241954.0 281317.6 262949.1 310255.7
#> 2002 258060.9 263689.3 252826.5 300768.6
#> 2003 269865.2 280474.5 282603.9 313818.2
#> 2004 289812.0 288337.3 291650.0 363988.2
#> 2005 308132.3 303469.5 321044.4 373969.8
#> 2006 332716.5 301337.7 309294.1 360086.1
#> 2007 305172.9 315596.6 329930.1 400535.5
#> 2008 312951.7 289190.7 301718.0 336997.0
#> 2009 294207.5 313871.7 318640.4 362644.2
#> 2010 341001.0 362288.2 354559.6 432789.9

# Denton PFD
cholette(s = xm, t = Y, rho = 1, lambda = 1)
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 129880.4 136947.1 154010.9 128260.9 149967.6 149425.7 149865.8 147018.4
#> 1993 117923.0 129325.2 153558.6 165712.9 155812.8 169672.7 170390.2 152481.9
#> 1994 146643.5 139322.0 160645.3 183188.9 180144.5 183755.4 192331.1 185913.5
#> 1995 168030.8 154432.3 174068.6 174259.2 185702.9 193770.3 208045.5 191091.3
#> 1996 162007.0 169247.2 189876.1 190232.7 209872.0 172304.2 223236.7 199777.7
#> 1997 178437.0 166379.2 194657.7 211151.8 214328.5 211860.5 192805.3 198307.4
#> 1998 192484.1 189127.0 206093.6 212838.5 218710.0 238269.3 228147.7 220925.1
#> 1999 189261.6 197361.0 219536.8 224324.7 241013.6 240608.8 230530.5 240000.7
#> 2000 211816.2 226203.1 251041.6 233897.9 256744.5 244441.9 244520.4 259336.5
#> 2001 212674.3 225142.9 246419.2 250874.6 285288.0 259384.0 229341.8 263487.1
#> 2002 233124.7 228537.6 258150.1 265095.5 271767.0 258043.2 272487.5 270125.0
#> 2003 247209.9 224120.7 260755.4 273534.2 280258.9 267102.7 268173.8 300994.3
#> 2004 254989.9 235768.1 301309.5 286874.8 296190.9 277129.4 288616.7 304571.4
#> 2005 272771.9 247382.7 318820.5 301917.3 317564.0 302970.9 312903.5 314171.8
#> 2006 279433.0 298454.5 339929.3 327638.0 351259.4 328456.1 314493.8 337023.6
#> 2007 305828.6 289972.4 348805.1 312910.1 370099.0 340167.7 336501.4 351612.6
#> 2008 304873.6 309143.0 344393.2 343060.5 364395.8 336798.7 360257.9 344479.4
#> 2009 270138.6 245865.6 320297.3 283285.8 307614.0 317917.7 299712.5 324247.3
#> 2010 293816.5 277293.7 327418.4 325090.5 334258.4 317337.0 342278.1 334673.6
#>           Sep      Oct      Nov      Dec
#> 1992 154882.0 177965.2 138568.2 198923.7
#> 1993 168995.1 185959.7 171584.7 200831.3
#> 1994 175373.9 160251.9 182457.1 219994.0
#> 1995 185424.4 182488.3 171245.5 233944.8
#> 1996 190708.5 212568.9 196774.4 250059.6
#> 1997 205414.6 225085.2 210866.8 264709.0
#> 1998 210683.1 210685.0 225139.6 234002.1
#> 1999 243071.2 226475.6 248796.7 307574.8
#> 2000 244435.9 253106.1 266869.2 296342.8
#> 2001 241694.7 280866.4 262493.3 310058.5
#> 2002 258106.5 263963.9 252851.7 302069.2
#> 2003 269671.6 280264.0 282360.2 313708.4
#> 2004 289800.4 288277.5 291600.9 365300.6
#> 2005 308046.8 303232.1 321274.9 375634.5
#> 2006 332771.1 301088.9 309247.9 360340.4
#> 2007 305366.2 315494.9 329490.9 399549.2
#> 2008 313196.7 289519.3 302985.5 339829.4
#> 2009 292288.6 310203.2 312478.4 354422.1
#> 2010 310753.7 331153.2 322026.0 400896.8

# Pro-rating
cholette(s = xm, t = Y, rho = 0, lambda = 0.5)
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 130111.6 137182.1 154255.0 128437.6 150133.7 149540.0 149918.0 146997.6
#> 1993 118411.7 129750.6 153938.2 165994.5 155964.7 169722.7 170332.8 152341.6
#> 1994 145860.1 138566.6 159800.6 182307.0 179414.9 183210.0 192030.9 185949.1
#> 1995 166171.0 153200.4 173168.3 173789.0 185598.4 194004.3 208584.4 191770.4
#> 1996 163108.6 170222.3 190766.0 190912.3 210379.5 172514.4 223233.7 199519.4
#> 1997 178761.7 166473.9 194570.8 210900.7 213977.2 211481.1 192487.3 198061.6
#> 1998 190028.4 187092.4 204302.5 211444.5 217762.5 237784.4 228227.5 221548.1
#> 1999 190262.1 198824.8 221460.2 226393.5 243129.1 242380.4 231680.9 240412.0
#> 2000 214740.9 227775.7 251390.1 233242.5 255269.1 242644.5 242637.6 257575.6
#> 2001 208625.7 222431.7 244952.6 250648.1 286156.8 260863.1 230984.8 265478.7
#> 2002 236359.4 230540.7 259286.0 265320.4 271253.5 257060.3 271135.5 268690.3
#> 2003 244283.0 222206.3 259310.5 272741.9 280081.2 267431.5 268901.2 302139.3
#> 2004 256733.0 237127.5 302715.2 287886.4 296887.0 277444.4 288584.6 304146.4
#> 2005 274049.4 248208.8 319498.9 302244.6 317627.0 302812.7 312563.6 313704.1
#> 2006 276755.6 295799.1 337231.5 325456.4 349477.9 327421.0 314206.2 337573.5
#> 2007 303556.4 288849.6 348497.4 313350.5 371231.1 341510.2 337890.3 352881.4
#> 2008 309147.3 312330.8 346783.7 344409.9 364862.0 336460.0 359193.4 342912.3
#> 2009 270206.6 245912.2 320339.4 283308.4 307624.1 317915.2 299699.5 324223.7
#> 2010 287409.1 271246.6 320278.3 318001.0 326969.0 310416.7 334813.9 327375.2
#>           Sep      Oct      Nov      Dec
#> 1992 154773.2 177727.1 138282.9 198357.3
#> 1993 168748.3 185596.6 171174.2 200272.1
#> 1994 175773.7 161003.8 183808.2 222296.2
#> 1995 186190.1 183277.8 171956.4 234793.3
#> 1996 190210.2 211724.8 195717.4 248356.4
#> 1997 205300.8 225179.4 211226.9 265581.6
#> 1998 211815.4 212372.8 227555.7 237170.9
#> 1999 242468.0 224758.7 245440.7 301345.6
#> 2000 243326.7 252852.8 267909.9 299390.5
#> 2001 243319.9 282205.9 262894.3 309163.3
#> 2002 256850.5 262997.9 252429.9 302397.5
#> 2003 270873.1 281584.0 283645.5 314956.5
#> 2004 289013.0 287102.9 290006.9 362782.6
#> 2005 307514.0 302682.4 320715.8 375069.8
#> 2006 334272.9 303417.1 312732.1 365792.7
#> 2007 306087.2 315648.4 328823.0 397472.5
#> 2008 311377.3 287560.3 300729.5 337166.4
#> 2009 292260.5 310167.7 312438.9 354374.9
#> 2010 303977.0 323931.5 315003.4 392154.3