rjd3bench: Temporal Disaggregation and Benchmarking in 'JDemetra+' 3.x

Interface to 'JDemetra+' 3.x (https://github.com/jdemetra) time series analysis software. It provides a variety of methods for temporal disaggregation & interpolation, benchmarking, reconciliation & multivariate temporal disaggregation and calendarization. It incorporates statistical methods described in the latest European Statistical System (ESS) guidelines on temporal disaggregation, benchmarking, and reconciliation (2018 edition). The package implements highly efficient algorithms for fast and reliable computation.

See also

Author

Maintainer: Corentin Lemasson corentin.lemasson@nbb.be

Authors:

Jean Palate palatejean@gmail.com

Other contributors:

Tanguy Barthelemy tanguy.barthelemy@insee.fr [contributor, artist]

Nikolina Rizanovska nikolina.rizanovska@gov.si [contributor]