Benchmarking by means of the Cholette method

Cholette method is based on a benchmarking methodology developed at Statistics Canada. It is a generalized model relying on the principle of movement preservation that encompasses several other benchmarking methods. The Denton method (both the AFD and PFD variants), as well as the naive pro-rating method, emerge as particular cases of the Cholette method. This method has been widely used for the purpose of benchmarking seasonally adjusted series among others.

Usage

cholette(
  s,
  t,
  rho = 1,
  lambda = 1,
  bias = c("None", "Additive", "Multiplicative"),
  conversion = c("Sum", "Average", "Last", "First", "UserDefined"),
  obsposition = 1L
)

Arguments

s

Preliminary series. Mandatory. It must be the same class as t.

t

Aggregation constraint. Mandatory. It must be either an object of class ts or a numeric vector.

rho

Numeric. Smoothing parameter whose value should be between 0 and 1. See vignette for more information on the choice of the rho parameter.

lambda

Numeric. Adjustment model parameter. Typically, lambda = 1 for proportional benchmarking; lambda = 0 for additive benchmarking; and lambda = 0.5 with rho = 0 for the naive pro-rating method. See vignette for more information on the choice of the lambda parameter.

bias

Character. Bias correction factor. No systematic bias is considered by default. See vignette for more details.

conversion

Conversion rule. Usually "Sum" or "Average". Sum by default.

obsposition

Position of the observation in the aggregated period (only used with "UserDefined" conversion).

Value

The benchmarked series is returned

Examples

ym_true <- rjd3toolkit::Retail$RetailSalesTotal
yq_true <- rjd3toolkit::aggregate(ym_true, 4)
Y_full <- rjd3toolkit::aggregate(ym_true, 1)

Y <- window(Y_full, end = c(2009,1)) # say no benchmark yet for the year 2010
xm <- ym_true + rnorm(n = length(ym_true), mean = -5000, sd = 10000)
xq <- rjd3toolkit::aggregate(xm, 4)

# Proportional benchmarking with a bias and some recommended value of rho for
# monthly and quarterly series respectively (see vignette)
cholette(s = xm, t = Y, rho = 0.9, lambda = 1, bias = "Multiplicative")
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 117913.3 136339.9 120087.5 151122.0 163599.9 168712.9 137979.0 154591.6
#> 1993 161350.8 119378.4 160992.6 141634.0 160472.7 164390.9 171957.7 155879.7
#> 1994 163950.1 144410.1 163585.0 169913.9 166349.7 162235.5 181679.1 183271.3
#> 1995 136812.5 150573.0 177569.7 186297.0 190104.0 184270.6 181432.5 190137.8
#> 1996 167640.6 153737.3 200606.4 189471.8 204302.3 197601.7 198216.8 211157.0
#> 1997 194276.6 173744.8 218786.0 199971.6 203019.6 198657.7 223013.8 216022.9
#> 1998 198074.5 182062.4 215999.1 211376.8 220235.4 212287.9 211543.4 207677.8
#> 1999 185370.7 205264.8 199583.3 221773.4 238867.7 230763.5 239795.5 250585.4
#> 2000 212433.8 240449.8 272546.6 240901.3 259518.5 252860.1 238884.1 254988.0
#> 2001 230405.0 234372.3 251541.7 234191.0 270508.5 254501.9 255415.7 285868.8
#> 2002 247395.1 228876.9 267259.5 257670.7 269046.7 257114.7 269703.6 272333.1
#> 2003 242010.5 219465.0 255193.2 271328.5 274081.6 269566.5 289957.4 287451.4
#> 2004 261667.0 265476.1 275068.8 287596.9 279045.9 302979.1 297159.1 293517.5
#> 2005 264468.0 263062.5 317456.8 313734.0 295830.4 331459.3 318161.5 328601.4
#> 2006 283835.9 287385.2 324715.8 305439.8 348565.7 338250.6 330860.5 346498.7
#> 2007 281249.3 275077.1 315681.8 333700.5 355223.4 354582.0 328547.1 371020.7
#> 2008 295106.2 298179.9 347920.9 328391.9 357521.5 347918.5 353987.2 346567.4
#> 2009 251067.5 267951.7 283160.8 290500.2 292491.9 340384.5 304301.6 318993.6
#> 2010 292549.5 287222.0 315401.6 336131.6 331003.3 327950.8 331009.1 333161.8
#>           Sep      Oct      Nov      Dec
#> 1992 149966.1 157743.1 153320.6 204340.0
#> 1993 165074.2 168519.0 157206.8 215391.0
#> 1994 176390.1 194010.2 184318.1 219907.9
#> 1995 186259.0 211681.0 193408.4 233958.3
#> 1996 167006.1 231364.6 202021.9 243538.5
#> 1997 208134.4 193828.1 192563.7 251983.8
#> 1998 219903.9 225355.2 218820.1 263768.4
#> 1999 241216.7 233455.1 247500.0 314380.0
#> 2000 230210.4 242400.4 250348.7 293214.2
#> 2001 237774.0 245538.0 261348.7 306259.4
#> 2002 238321.6 259708.9 250258.6 316632.5
#> 2003 264596.9 266179.2 291912.9 336410.9
#> 2004 269827.8 280681.5 304493.4 362916.9
#> 2005 285634.1 297340.4 302999.1 377943.4
#> 2006 302570.3 296564.0 316379.5 399070.1
#> 2007 323827.2 338444.8 348977.0 379467.1
#> 2008 307577.7 324679.3 298888.8 346193.5
#> 2009 303220.3 295028.3 317403.8 373966.8
#> 2010 305823.3 336810.7 341256.9 393591.2
cholette(s = xq, t = Y, rho = 0.729, lambda = 1, bias = "Multiplicative")
#>           Qtr1      Qtr2      Qtr3      Qtr4
#> 1992  374268.3  483380.9  442461.7  515605.1
#> 1993  441437.9  466461.8  492965.7  541382.7
#> 1994  471955.4  498515.2  541347.5  598202.8
#> 1995  464948.9  560640.7  557814.6  639099.8
#> 1996  522116.3  591544.1  576445.6  676559.0
#> 1997  586664.0  601298.9  646980.7  639059.4
#> 1998  595756.3  643962.7  639419.0  707967.0
#> 1999  590843.2  691663.9  731480.0  794569.0
#> 2000  725293.9  753116.4  723980.5  786365.2
#> 2001  715929.0  759154.1  779173.5  813468.4
#> 2002  743823.0  784032.0  780372.7  826094.3
#> 2003  716631.5  814762.0  841860.1  894900.5
#> 2004  802118.3  869864.7  860761.4  947685.6
#> 2005  845174.0  940853.2  932117.9  978545.9
#> 2006  895589.6  992310.3  980091.5 1012144.6
#> 2007  872143.5 1043408.7 1023276.8 1066969.0
#> 2008  941231.0 1033998.2 1008298.4  969405.5
#> 2009  802423.8  923341.8  926367.8  986337.6
#> 2010  894520.6  994524.3  969583.0 1071368.1

# Proportional benchmarking with no bias
xm_no_bias <- ym_true + rnorm(n = length(ym_true), mean = 0, sd = 10000)
cholette(s = xm_no_bias, t = Y, rho = 0.9, lambda = 1)
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 141379.4 119261.9 149114.4 138125.8 164717.7 140281.9 157659.3 149982.5
#> 1993 128835.7 147845.7 163487.3 169527.6 147922.4 174818.3 162655.7 161225.6
#> 1994 133137.1 139120.5 189096.6 163813.6 180340.7 175955.3 169636.8 200226.7
#> 1995 151884.1 136824.5 204682.9 170085.6 192580.3 197131.7 186132.9 173705.2
#> 1996 187611.3 179029.5 196561.1 185470.7 211552.2 183590.0 198114.7 202624.0
#> 1997 191959.2 172143.4 199567.9 195006.4 199723.0 208813.4 223558.4 195871.0
#> 1998 183221.9 172409.5 217654.1 222868.8 221906.5 215655.7 222583.2 234997.3
#> 1999 201794.5 180889.6 233716.7 227130.2 248762.8 233746.6 225779.7 241524.9
#> 2000 231858.0 209157.3 253448.3 248416.9 256317.4 256742.5 250341.1 233662.4
#> 2001 214645.2 212644.3 255221.9 258812.9 268215.6 283567.0 252731.7 254860.0
#> 2002 228541.4 234246.3 248563.3 252273.8 272565.9 273034.5 276364.5 287702.4
#> 2003 235607.9 241273.7 269974.5 281700.5 278614.7 279633.3 261242.1 276216.8
#> 2004 244014.0 244936.8 285880.7 297298.1 301798.5 286323.4 294763.3 307396.3
#> 2005 265833.5 263493.0 303528.6 307906.7 300803.4 316692.4 308712.4 326504.9
#> 2006 271089.6 292770.6 323351.2 312488.9 354930.9 322969.9 317160.0 350549.4
#> 2007 291170.7 275838.7 338247.3 317771.5 349123.4 336085.7 349781.0 358434.2
#> 2008 316896.8 310102.9 346106.8 325317.9 332582.7 351907.9 342322.2 328624.9
#> 2009 285290.1 281781.4 288592.5 293198.0 302262.8 308005.3 316783.8 303521.1
#> 2010 290718.8 279089.8 315896.8 314753.2 348313.5 330963.8 347252.8 334553.7
#>           Sep      Oct      Nov      Dec
#> 1992 143125.1 166180.5 169147.3 176740.1
#> 1993 161833.4 163576.8 160377.5 200142.0
#> 1994 179721.6 167409.4 187742.6 223820.3
#> 1995 183088.4 186551.0 193950.8 245886.7
#> 1996 188574.7 208392.1 190086.7 235057.9
#> 1997 186523.1 218124.6 213077.5 269635.2
#> 1998 226523.0 203638.3 212633.8 253013.0
#> 1999 235970.6 235077.9 253486.1 290676.5
#> 2000 251720.3 240203.3 252328.0 304560.5
#> 2001 226611.5 284006.8 262088.7 294319.4
#> 2002 238245.8 273426.0 253859.9 295498.0
#> 2003 275973.0 272023.5 274462.7 321431.5
#> 2004 286920.2 284222.3 293003.5 353872.8
#> 2005 314904.4 291839.8 308678.2 387793.8
#> 2006 297132.1 302648.7 331249.7 403795.1
#> 2007 325098.9 316417.4 354110.0 393719.3
#> 2008 319291.3 325779.3 309407.7 344592.7
#> 2009 280052.5 314201.3 297906.2 366875.8
#> 2010 305930.4 307278.8 335709.7 375194.5

# Additive benchmarking
cholette(s = xm, t = Y, rho = 0.9, lambda = 0, bias = "Additive")
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 130292.9 144477.8 126272.5 153138.9 162857.7 166042.6 135929.2 150697.5
#> 1993 158362.9 119356.6 159705.6 141862.3 160410.6 164610.7 172235.5 157007.8
#> 1994 164660.2 145453.5 163989.1 170054.2 166483.1 162389.3 181387.0 182931.9
#> 1995 137524.1 151085.7 177620.1 186225.1 189998.0 184314.8 181565.4 190122.9
#> 1996 168205.9 154646.5 200464.4 189616.7 204109.4 197605.9 198239.7 210855.2
#> 1997 194798.6 175110.5 218409.5 200327.8 203225.8 198980.1 222370.8 215592.6
#> 1998 197666.1 181639.9 215608.0 211061.5 220114.8 212167.3 211496.9 207649.2
#> 1999 185467.0 205526.8 199887.8 222069.1 239058.2 231005.4 239929.7 250542.0
#> 2000 213614.2 240762.1 271683.2 241237.7 259125.4 252733.8 239317.5 254769.2
#> 2001 230790.2 234612.6 251415.2 234392.4 270176.9 254438.4 255375.2 285598.1
#> 2002 247419.1 228942.7 267206.5 257652.5 269000.6 257124.2 269674.1 272299.9
#> 2003 242547.4 220459.0 255600.8 271419.7 274097.1 269655.8 289584.2 287113.5
#> 2004 261687.0 265453.9 274983.7 287462.1 279053.2 302877.0 297169.7 293618.7
#> 2005 265453.0 264176.9 317323.5 313651.1 296196.2 330822.1 317868.4 328012.9
#> 2006 284180.9 287645.4 324519.5 305506.3 348326.7 338160.6 330851.0 346447.0
#> 2007 281162.1 274946.3 315449.7 333436.0 354948.3 354358.2 328431.6 370953.5
#> 2008 295932.0 299158.7 349037.8 329468.6 358490.0 348650.4 354357.7 346497.0
#> 2009 248911.5 265055.4 279766.4 287062.8 289461.6 336634.2 302827.5 318526.1
#> 2010 306577.4 304693.3 335031.5 357887.0 355332.6 354549.1 359503.1 363380.5
#>           Sep      Oct      Nov      Dec
#> 1992 145862.2 153056.3 149128.0 197960.2
#> 1993 166026.8 169425.5 158447.0 214796.7
#> 1994 176186.0 193487.1 184014.3 218985.3
#> 1995 186346.7 211257.2 193381.8 233062.2
#> 1996 168038.7 230497.8 202123.0 242261.7
#> 1997 207909.2 193914.6 192526.9 250836.5
#> 1998 220159.2 225757.7 219160.4 264623.9
#> 1999 241326.4 233753.0 247498.1 312492.6
#> 2000 230926.3 242608.2 250231.4 291747.0
#> 2001 237899.6 245608.8 261342.0 306075.7
#> 2002 238538.2 259813.6 250511.1 316139.4
#> 2003 264716.8 266233.2 291464.4 335262.1
#> 2004 270218.7 281056.9 304654.5 362194.6
#> 2005 286188.7 297554.2 303038.6 376405.5
#> 2006 302649.0 296620.8 316360.0 398868.8
#> 2007 323927.3 338682.6 349393.2 380109.2
#> 2008 307189.5 323564.5 297369.7 343216.9
#> 2009 305156.3 299509.7 323828.3 381731.2
#> 2010 338277.1 369987.1 375633.1 428002.8

# Denton PFD
cholette(s = xm, t = Y, rho = 1, lambda = 1)
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 119207.4 137418.8 120713.4 151537.7 163708.8 168544.6 137671.5 154097.2
#> 1993 160972.1 119210.8 160879.7 141609.3 160502.5 164459.9 172053.0 155976.3
#> 1994 164036.2 144480.9 163651.1 169962.7 166375.7 162239.2 181660.1 183232.8
#> 1995 136783.3 150552.7 177558.3 186296.4 190112.9 184286.1 181451.8 190158.7
#> 1996 167598.2 153685.7 200530.1 189402.5 204239.6 197561.3 198202.2 211172.4
#> 1997 194382.3 173838.7 218890.9 200050.4 203073.1 198676.6 222990.6 215960.5
#> 1998 198103.4 182134.1 216109.2 211480.8 220322.3 212338.9 211558.4 207658.3
#> 1999 185176.7 205029.8 199369.2 221580.8 238730.0 230708.6 239823.1 250697.0
#> 2000 212549.2 240521.8 272579.9 240906.1 259502.5 252831.1 238848.0 254941.2
#> 2001 230492.7 234499.7 251686.2 234309.0 270608.1 254549.2 255412.2 285809.3
#> 2002 247256.8 228763.8 267153.2 257595.7 269001.1 257103.7 269727.6 272390.6
#> 2003 242071.2 219501.1 255222.0 271350.7 274097.6 269575.4 289957.2 287439.3
#> 2004 261640.8 265455.4 275048.4 287575.4 279026.4 302964.0 297153.3 293523.7
#> 2005 264459.0 263038.2 317422.1 313705.2 295812.6 331448.3 318159.4 328605.9
#> 2006 283936.4 287506.7 324849.8 305542.6 348644.9 338280.6 330841.6 346432.4
#> 2007 281158.1 275006.3 315628.3 333672.6 355220.3 354597.5 328571.9 371059.3
#> 2008 295143.1 298222.1 347984.2 328462.4 357610.0 348007.1 354068.9 346621.6
#> 2009 250681.9 267442.4 282582.0 289923.2 291980.4 339923.6 304066.7 318977.2
#> 2010 294210.6 289238.9 317998.8 339267.2 334417.6 331625.3 334983.4 337402.8
#>           Sep      Oct      Nov      Dec
#> 1992 149393.7 157095.7 152706.6 203620.6
#> 1993 165176.6 168617.9 157292.0 215497.8
#> 1994 176339.3 193945.6 184255.9 219841.6
#> 1995 186276.2 211691.8 193405.0 233930.7
#> 1996 167044.4 231446.1 202115.7 243666.9
#> 1997 208043.3 193727.3 192467.1 251902.0
#> 1998 219851.8 225269.6 218701.8 263576.4
#> 1999 241389.6 233664.7 247737.0 314649.6
#> 2000 230167.3 242358.3 250321.9 293228.7
#> 2001 237682.0 245410.3 261193.0 306073.4
#> 2002 238395.4 259810.4 250365.0 316758.6
#> 2003 264573.8 266144.1 291865.2 336356.4
#> 2004 269845.5 280712.0 304532.8 362952.3
#> 2005 285644.5 297357.4 303028.2 378010.2
#> 2006 302476.9 296450.7 316252.2 398921.4
#> 2007 323862.0 338484.1 349020.4 379517.1
#> 2008 307582.7 324617.8 298740.0 345873.0
#> 2009 303467.3 295561.7 318331.4 375533.3
#> 2010 309915.4 341515.2 346203.9 399484.4

# Pro-rating
cholette(s = xm, t = Y, rho = 0, lambda = 0.5)
#>           Jan      Feb      Mar      Apr      May      Jun      Jul      Aug
#> 1992 118765.9 136926.5 120312.8 151093.3 163317.5 168261.6 137562.4 154133.9
#> 1993 159233.3 118145.9 159743.3 140873.7 159968.7 164220.1 172123.3 156330.8
#> 1994 162951.3 143766.8 163093.9 169619.3 166243.3 162282.9 181874.3 183583.4
#> 1995 136983.3 150741.2 177741.6 186445.6 190219.0 184341.9 181458.3 190112.7
#> 1996 168565.0 154498.2 201468.6 190143.3 204850.0 197937.2 198332.9 211016.1
#> 1997 193432.4 172724.0 217287.2 198555.2 201667.1 197551.9 222166.6 215759.5
#> 1998 194488.8 179887.9 214581.6 210927.4 220546.0 213139.4 212756.3 209046.4
#> 1999 189002.7 208514.5 202036.2 223753.0 240227.9 231353.0 239668.9 249685.8
#> 2000 213461.4 240897.1 272416.2 240393.8 258699.0 251953.6 238067.6 254299.5
#> 2001 227082.7 231839.3 249643.1 233105.4 269961.8 254570.4 255997.2 287017.7
#> 2002 247994.9 229504.1 268029.1 258388.4 269711.7 257608.1 270010.8 272362.1
#> 2003 243037.2 219941.8 255312.3 271100.3 273603.7 268960.1 289269.5 286852.8
#> 2004 260901.9 265246.3 275275.8 288150.6 279781.7 303860.7 297961.2 294108.8
#> 2005 265941.7 263803.4 317633.1 313378.2 295154.4 330481.4 317185.2 327724.5
#> 2006 280817.5 285152.7 323026.8 304536.8 348220.8 338475.1 331532.2 347584.1
#> 2007 281057.0 274873.5 315451.9 333478.7 355026.3 354437.9 328475.5 371029.5
#> 2008 296531.2 299692.3 349675.9 329923.5 358937.3 348920.1 354492.5 346423.2
#> 2009 253809.7 269951.5 284432.9 291080.1 292478.2 339818.8 303455.3 317880.6
#> 2010 283582.1 278790.1 306511.0 327011.1 322336.6 319645.3 322882.0 325214.0
#>           Sep      Oct      Nov      Dec
#> 1992 149607.4 157532.6 153363.2 204839.0
#> 1993 165859.8 169629.3 158528.0 217591.8
#> 1994 176774.8 194499.0 184817.4 220514.6
#> 1995 186177.1 211515.3 193182.7 233585.2
#> 1996 166659.3 230520.8 200929.6 241744.1
#> 1997 208589.0 195075.0 194784.7 256410.4
#> 1998 221357.0 226643.0 219668.0 264063.4
#> 1999 239608.4 231169.2 244284.0 309252.3
#> 2000 229895.7 242526.0 251106.5 295039.5
#> 2001 239074.6 247185.4 263371.2 308876.3
#> 2002 238037.5 259002.9 249127.9 314544.5
#> 2003 264232.0 266102.4 292265.1 337476.8
#> 2004 270059.4 280475.9 303643.4 360964.3
#> 2005 285141.9 297249.9 303493.8 379503.4
#> 2006 303768.4 297920.9 317959.0 401141.5
#> 2007 323925.6 338663.1 349341.5 380037.5
#> 2008 306761.1 322976.0 296426.8 342173.1
#> 2009 302077.6 293950.7 316402.1 373133.5
#> 2010 298719.7 329177.9 333697.2 385053.0